requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
Flask==1.1.2
PyYAML==5.3.1
flask-cors
pyopenssl
timeloop

